Social Nav Buttons to Separate Line
Moved social nav buttons to a separate “nav-wrapper” tag so that they
show up under the main site nav buttons
    - current main site nav buttons are “Blog” and “Blog2” both linking
to the same place…blog

Hoping to have these maybe connected to tags in some way? Such that
what is revealed with the “panel” moves away is just a selection of
certain posts…

Would need to figure out how to handle the transition (if I want one)
from say “blog” to “professional” or whatever other sections.